ASSIGNMENT: DECORATING OUR PARKING CHARGES
