--- ParkingManagement/src/main/java/edu/du/ict4315/parkingmanagement/factory/strategy/charges/ParkingChargeStrategy.java
@@ -1,22 +1,0 @@
-/**
- * @Course: ICT 4315
- * @Project: ParkingManagement
- * @Instructor: Mike Prasad
- */
-package edu.du.ict4315.parkingmanagement.factory.strategy.charges;
-
-import edu.du.ict4315.parkingmanagement.Money;
-import edu.du.ict4315.parkingmanagement.VehicleType;
-
-import java.time.LocalDate;
-
-/**
- *
- * @Date: 1/19/23
- * @author lutherchikumba
- *
- */
-public interface ParkingChargeStrategy {
-      String  getStrategy();
-      Money getDiscountRate();
-}
